+++ CircleciScripts/check_sdk_on_maven.py
@@ -81,14 +81,14 @@ def IsAllPackageAvailableOnMaven(version):
         return (True, not_available_packagelist)
 
 version = sys.argv[1]
+email_from = sys.argv[2]
+email_to = sys.argv[3]
 starttime = datetime.datetime.now()
 ready, not_available_packagelist = IsAllPackageAvailableOnMaven(version)
 if not ready:
     # send email 
-    email_from = "aws-mobile-sdk-dev@amazon.com"
-    email_to = "aws-mobile-sdk-dev@amazon.com"
     destination ='{{ "ToAddresses":  ["{0}"] }}'.format(email_to)
-    title = "Android SDK on Maven borken"
+    title = "The latest AWS Android SDKs are missing on Maven"
     body = "Android SDKs {0} are still not available on maven. Below packages are not found on maven: {1}".format(version, not_available_packagelist)
     message = '{{"Subject": {{"Data": "{0}","Charset": "UTF-8" }},"Body": {{ "Text": {{ "Data": "{1}", "Charset": "UTF-8" }} }} }}'.format(title, body)
     profile = 'circleci-info'